Documents/GitHub/Mobility_Gab/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-environ==0.11.2
django-filter==23.5
drf-yasg==1.21.7
psycopg[binary]==3.1.18
Pillow==10.2.0
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1


